Crate sui_protocol_config

Macros

check_limit

check_limit_by_meter

Used to check which limits were crossed if the TX is metered (not system tx) Args are: is_metered, value_to_check, metered_limit, unmetered_limit metered_limit is always less than or equal to unmetered_hard_limit

Structs

Error

OverrideGuard

ProtocolConfig

Constants that change the behavior of the protocol.

ProtocolVersion

SupportedProtocolVersions

Models the set of protocol versions supported by a validator. The sui-node binary will always use the SYSTEM_DEFAULT constant, but for testing we need to be able to inject arbitrary versions into SuiNode.

Enums

Chain

ConsensusTransactionOrdering

Ordering mechanism for transactions in one Narwhal consensus output.

LimitThresholdCrossed

Defines which limit got crossed. The value which crossed the limit and value of the limit crossed are embedded

ProtocolConfigValue

Functions

check_limit_in_range

Convenience function for comparing limit ranges V::MAX must be at >= U::MAX and T::MAX