Struct sui_protocol_config::ProtocolConfig

pub struct ProtocolConfig {
    pub version: ProtocolVersion,
    /* private fields */
}

Constants that change the behavior of the protocol.

The value of each constant here must be fixed for a given protocol version. To change the value of a constant, advance the protocol version, and add support for it in get_for_version under the new version number. (below).

To add a new field to this struct, use the following procedure:

Advance the protocol version.

Add the field as a private Option<T> to the struct.

Initialize the field to None in prior protocol versions.

Initialize the field to Some(val) for your new protocol version.

Add a public getter that simply unwraps the field.

A public getter of the form field(&self) -> field_type will be automatically generated for you. Example for a field: new_constant: Option<u64>

     pub fn new_constant(&self) -> u64 {
        self.new_constant.expect(Self::CONSTANT_ERR_MSG)
    }

This way, if the constant is accessed in a protocol version in which it is not defined, the validator will crash. (Crashing is necessary because this type of error would almost always result in forking if not prevented here).

If you want a customized getter, you can add a method in the impl. For example see max_size_written_objects_as_option

Fields
version: ProtocolVersion
Implementations

impl ProtocolConfig

pub fn max_tx_size_bytes(&self) -> u64

pub fn max_tx_size_bytes_as_option(&self) -> Option<u64>

pub fn max_input_objects(&self) -> u64

pub fn max_input_objects_as_option(&self) -> Option<u64>

pub fn max_size_written_objects(&self) -> u64

pub fn max_size_written_objects_as_option(&self) -> Option<u64>

pub fn max_size_written_objects_system_tx(&self) -> u64

pub fn max_size_written_objects_system_tx_as_option(&self) -> Option<u64>

pub fn max_serialized_tx_effects_size_bytes(&self) -> u64

pub fn max_serialized_tx_effects_size_bytes_as_option(&self) -> Option<u64>

pub fn max_serialized_tx_effects_size_bytes_system_tx(&self) -> u64

pub fn max_serialized_tx_effects_size_bytes_system_tx_as_option( &self ) -> Option<u64>

pub fn max_gas_payment_objects(&self) -> u32

pub fn max_gas_payment_objects_as_option(&self) -> Option<u32>

pub fn max_modules_in_publish(&self) -> u32

pub fn max_modules_in_publish_as_option(&self) -> Option<u32>

pub fn max_arguments(&self) -> u32

pub fn max_arguments_as_option(&self) -> Option<u32>

pub fn max_type_arguments(&self) -> u32

pub fn max_type_arguments_as_option(&self) -> Option<u32>

pub fn max_type_argument_depth(&self) -> u32

pub fn max_type_argument_depth_as_option(&self) -> Option<u32>

pub fn max_pure_argument_size(&self) -> u32

pub fn max_pure_argument_size_as_option(&self) -> Option<u32>

pub fn max_programmable_tx_commands(&self) -> u32

pub fn max_programmable_tx_commands_as_option(&self) -> Option<u32>

pub fn move_binary_format_version(&self) -> u32

pub fn move_binary_format_version_as_option(&self) -> Option<u32>

pub fn max_move_object_size(&self) -> u64

pub fn max_move_object_size_as_option(&self) -> Option<u64>

pub fn max_move_package_size(&self) -> u64

pub fn max_move_package_size_as_option(&self) -> Option<u64>

pub fn max_tx_gas(&self) -> u64

pub fn max_tx_gas_as_option(&self) -> Option<u64>

pub fn max_gas_price(&self) -> u64

pub fn max_gas_price_as_option(&self) -> Option<u64>

pub fn max_gas_computation_bucket(&self) -> u64

pub fn max_gas_computation_bucket_as_option(&self) -> Option<u64>

pub fn max_loop_depth(&self) -> u64

pub fn max_loop_depth_as_option(&self) -> Option<u64>

pub fn max_generic_instantiation_length(&self) -> u64

pub fn max_generic_instantiation_length_as_option(&self) -> Option<u64>

pub fn max_function_parameters(&self) -> u64

pub fn max_function_parameters_as_option(&self) -> Option<u64>

pub fn max_basic_blocks(&self) -> u64

pub fn max_basic_blocks_as_option(&self) -> Option<u64>

pub fn max_value_stack_size(&self) -> u64

pub fn max_value_stack_size_as_option(&self) -> Option<u64>

pub fn max_type_nodes(&self) -> u64

pub fn max_type_nodes_as_option(&self) -> Option<u64>

pub fn max_push_size(&self) -> u64

pub fn max_push_size_as_option(&self) -> Option<u64>

pub fn max_struct_definitions(&self) -> u64

pub fn max_struct_definitions_as_option(&self) -> Option<u64>

pub fn max_function_definitions(&self) -> u64

pub fn max_function_definitions_as_option(&self) -> Option<u64>

pub fn max_fields_in_struct(&self) -> u64

pub fn max_fields_in_struct_as_option(&self) -> Option<u64>

pub fn max_dependency_depth(&self) -> u64

pub fn max_dependency_depth_as_option(&self) -> Option<u64>

pub fn max_num_event_emit(&self) -> u64

pub fn max_num_event_emit_as_option(&self) -> Option<u64>

pub fn max_num_new_move_object_ids(&self) -> u64

pub fn max_num_new_move_object_ids_as_option(&self) -> Option<u64>

pub fn max_num_new_move_object_ids_system_tx(&self) -> u64

pub fn max_num_new_move_object_ids_system_tx_as_option(&self) -> Option<u64>

pub fn max_num_deleted_move_object_ids(&self) -> u64

pub fn max_num_deleted_move_object_ids_as_option(&self) -> Option<u64>

pub fn max_num_deleted_move_object_ids_system_tx(&self) -> u64

pub fn max_num_deleted_move_object_ids_system_tx_as_option(&self) -> Option<u64>

pub fn max_num_transferred_move_object_ids(&self) -> u64

pub fn max_num_transferred_move_object_ids_as_option(&self) -> Option<u64>

pub fn max_num_transferred_move_object_ids_system_tx(&self) -> u64

pub fn max_num_transferred_move_object_ids_system_tx_as_option( &self ) -> Option<u64>

pub fn max_event_emit_size(&self) -> u64

pub fn max_event_emit_size_as_option(&self) -> Option<u64>

pub fn max_move_vector_len(&self) -> u64

pub fn max_move_vector_len_as_option(&self) -> Option<u64>

pub fn max_move_identifier_len(&self) -> u64

pub fn max_move_identifier_len_as_option(&self) -> Option<u64>

pub fn max_move_value_depth(&self) -> u64

pub fn max_move_value_depth_as_option(&self) -> Option<u64>

pub fn max_back_edges_per_function(&self) -> u64

pub fn max_back_edges_per_function_as_option(&self) -> Option<u64>

pub fn max_back_edges_per_module(&self) -> u64

pub fn max_back_edges_per_module_as_option(&self) -> Option<u64>

pub fn max_verifier_meter_ticks_per_function(&self) -> u64

pub fn max_verifier_meter_ticks_per_function_as_option(&self) -> Option<u64>

pub fn max_meter_ticks_per_module(&self) -> u64

pub fn max_meter_ticks_per_module_as_option(&self) -> Option<u64>

pub fn object_runtime_max_num_cached_objects(&self) -> u64

pub fn object_runtime_max_num_cached_objects_as_option(&self) -> Option<u64>

pub fn object_runtime_max_num_cached_objects_system_tx(&self) -> u64

pub fn object_runtime_max_num_cached_objects_system_tx_as_option( &self ) -> Option<u64>

pub fn object_runtime_max_num_store_entries(&self) -> u64

pub fn object_runtime_max_num_store_entries_as_option(&self) -> Option<u64>

pub fn object_runtime_max_num_store_entries_system_tx(&self) -> u64

pub fn object_runtime_max_num_store_entries_system_tx_as_option( &self ) -> Option<u64>

pub fn base_tx_cost_fixed(&self) -> u64

pub fn base_tx_cost_fixed_as_option(&self) -> Option<u64>

pub fn package_publish_cost_fixed(&self) -> u64

pub fn package_publish_cost_fixed_as_option(&self) -> Option<u64>

pub fn base_tx_cost_per_byte(&self) -> u64

pub fn base_tx_cost_per_byte_as_option(&self) -> Option<u64>

pub fn package_publish_cost_per_byte(&self) -> u64

pub fn package_publish_cost_per_byte_as_option(&self) -> Option<u64>

pub fn obj_access_cost_read_per_byte(&self) -> u64

pub fn obj_access_cost_read_per_byte_as_option(&self) -> Option<u64>

pub fn obj_access_cost_mutate_per_byte(&self) -> u64

pub fn obj_access_cost_mutate_per_byte_as_option(&self) -> Option<u64>

pub fn obj_access_cost_delete_per_byte(&self) -> u64

pub fn obj_access_cost_delete_per_byte_as_option(&self) -> Option<u64>

pub fn obj_access_cost_verify_per_byte(&self) -> u64

pub fn obj_access_cost_verify_per_byte_as_option(&self) -> Option<u64>

pub fn gas_model_version(&self) -> u64

pub fn gas_model_version_as_option(&self) -> Option<u64>

pub fn obj_data_cost_refundable(&self) -> u64

pub fn obj_data_cost_refundable_as_option(&self) -> Option<u64>

pub fn obj_metadata_cost_non_refundable(&self) -> u64

pub fn obj_metadata_cost_non_refundable_as_option(&self) -> Option<u64>

pub fn storage_rebate_rate(&self) -> u64

pub fn storage_rebate_rate_as_option(&self) -> Option<u64>

pub fn storage_fund_reinvest_rate(&self) -> u64

pub fn storage_fund_reinvest_rate_as_option(&self) -> Option<u64>

pub fn reward_slashing_rate(&self) -> u64

pub fn reward_slashing_rate_as_option(&self) -> Option<u64>

pub fn storage_gas_price(&self) -> u64

pub fn storage_gas_price_as_option(&self) -> Option<u64>

pub fn max_transactions_per_checkpoint(&self) -> u64

pub fn max_transactions_per_checkpoint_as_option(&self) -> Option<u64>

pub fn max_checkpoint_size_bytes(&self) -> u64

pub fn max_checkpoint_size_bytes_as_option(&self) -> Option<u64>

pub fn buffer_stake_for_protocol_upgrade_bps(&self) -> u64

pub fn buffer_stake_for_protocol_upgrade_bps_as_option(&self) -> Option<u64>

pub fn address_from_bytes_cost_base(&self) -> u64

pub fn address_from_bytes_cost_base_as_option(&self) -> Option<u64>

pub fn address_to_u256_cost_base(&self) -> u64

pub fn address_to_u256_cost_base_as_option(&self) -> Option<u64>

pub fn address_from_u256_cost_base(&self) -> u64

pub fn address_from_u256_cost_base_as_option(&self) -> Option<u64>

pub fn dynamic_field_hash_type_and_key_cost_base(&self) -> u64

pub fn dynamic_field_hash_type_and_key_cost_base_as_option(&self) -> Option<u64>

pub fn dynamic_field_hash_type_and_key_type_cost_per_byte(&self) -> u64

pub fn dynamic_field_hash_type_and_key_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_hash_type_and_key_value_cost_per_byte(&self) -> u64

pub fn dynamic_field_hash_type_and_key_value_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_hash_type_and_key_type_tag_cost_per_byte(&self) -> u64

pub fn dynamic_field_hash_type_and_key_type_tag_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_add_child_object_cost_base(&self) -> u64

pub fn dynamic_field_add_child_object_cost_base_as_option(&self) -> Option<u64>

pub fn dynamic_field_add_child_object_type_cost_per_byte(&self) -> u64

pub fn dynamic_field_add_child_object_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_add_child_object_value_cost_per_byte(&self) -> u64

pub fn dynamic_field_add_child_object_value_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_add_child_object_struct_tag_cost_per_byte(&self) -> u64

pub fn dynamic_field_add_child_object_struct_tag_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_borrow_child_object_cost_base(&self) -> u64

pub fn dynamic_field_borrow_child_object_cost_base_as_option( &self ) -> Option<u64>

pub fn dynamic_field_borrow_child_object_child_ref_cost_per_byte(&self) -> u64

pub fn dynamic_field_borrow_child_object_child_ref_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_borrow_child_object_type_cost_per_byte(&self) -> u64

pub fn dynamic_field_borrow_child_object_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_remove_child_object_cost_base(&self) -> u64

pub fn dynamic_field_remove_child_object_cost_base_as_option( &self ) -> Option<u64>

pub fn dynamic_field_remove_child_object_child_cost_per_byte(&self) -> u64

pub fn dynamic_field_remove_child_object_child_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_remove_child_object_type_cost_per_byte(&self) -> u64

pub fn dynamic_field_remove_child_object_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_has_child_object_cost_base(&self) -> u64

pub fn dynamic_field_has_child_object_cost_base_as_option(&self) -> Option<u64>

pub fn dynamic_field_has_child_object_with_ty_cost_base(&self) -> u64

pub fn dynamic_field_has_child_object_with_ty_cost_base_as_option( &self ) -> Option<u64>

pub fn dynamic_field_has_child_object_with_ty_type_cost_per_byte(&self) -> u64

pub fn dynamic_field_has_child_object_with_ty_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn dynamic_field_has_child_object_with_ty_type_tag_cost_per_byte( &self ) -> u64

pub fn dynamic_field_has_child_object_with_ty_type_tag_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn event_emit_cost_base(&self) -> u64

pub fn event_emit_cost_base_as_option(&self) -> Option<u64>

pub fn event_emit_value_size_derivation_cost_per_byte(&self) -> u64

pub fn event_emit_value_size_derivation_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn event_emit_tag_size_derivation_cost_per_byte(&self) -> u64

pub fn event_emit_tag_size_derivation_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn event_emit_output_cost_per_byte(&self) -> u64

pub fn event_emit_output_cost_per_byte_as_option(&self) -> Option<u64>

pub fn object_borrow_uid_cost_base(&self) -> u64

pub fn object_borrow_uid_cost_base_as_option(&self) -> Option<u64>

pub fn object_delete_impl_cost_base(&self) -> u64

pub fn object_delete_impl_cost_base_as_option(&self) -> Option<u64>

pub fn object_record_new_uid_cost_base(&self) -> u64

pub fn object_record_new_uid_cost_base_as_option(&self) -> Option<u64>

pub fn transfer_transfer_internal_cost_base(&self) -> u64

pub fn transfer_transfer_internal_cost_base_as_option(&self) -> Option<u64>

pub fn transfer_freeze_object_cost_base(&self) -> u64

pub fn transfer_freeze_object_cost_base_as_option(&self) -> Option<u64>

pub fn transfer_share_object_cost_base(&self) -> u64

pub fn transfer_share_object_cost_base_as_option(&self) -> Option<u64>

pub fn tx_context_derive_id_cost_base(&self) -> u64

pub fn tx_context_derive_id_cost_base_as_option(&self) -> Option<u64>

pub fn types_is_one_time_witness_cost_base(&self) -> u64

pub fn types_is_one_time_witness_cost_base_as_option(&self) -> Option<u64>

pub fn types_is_one_time_witness_type_tag_cost_per_byte(&self) -> u64

pub fn types_is_one_time_witness_type_tag_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn types_is_one_time_witness_type_cost_per_byte(&self) -> u64

pub fn types_is_one_time_witness_type_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn validator_validate_metadata_cost_base(&self) -> u64

pub fn validator_validate_metadata_cost_base_as_option(&self) -> Option<u64>

pub fn validator_validate_metadata_data_cost_per_byte(&self) -> u64

pub fn validator_validate_metadata_data_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn crypto_invalid_arguments_cost(&self) -> u64

pub fn crypto_invalid_arguments_cost_as_option(&self) -> Option<u64>

pub fn bls12381_bls12381_min_sig_verify_cost_base(&self) -> u64

pub fn bls12381_bls12381_min_sig_verify_cost_base_as_option( &self ) -> Option<u64>

pub fn bls12381_bls12381_min_sig_verify_msg_cost_per_byte(&self) -> u64

pub fn bls12381_bls12381_min_sig_verify_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn bls12381_bls12381_min_sig_verify_msg_cost_per_block(&self) -> u64

pub fn bls12381_bls12381_min_sig_verify_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn bls12381_bls12381_min_pk_verify_cost_base(&self) -> u64

pub fn bls12381_bls12381_min_pk_verify_cost_base_as_option(&self) -> Option<u64>

pub fn bls12381_bls12381_min_pk_verify_msg_cost_per_byte(&self) -> u64

pub fn bls12381_bls12381_min_pk_verify_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn bls12381_bls12381_min_pk_verify_msg_cost_per_block(&self) -> u64

pub fn bls12381_bls12381_min_pk_verify_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_ecrecover_keccak256_cost_base(&self) -> u64

pub fn ecdsa_k1_ecrecover_keccak256_cost_base_as_option(&self) -> Option<u64>

pub fn ecdsa_k1_ecrecover_keccak256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_k1_ecrecover_keccak256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_ecrecover_keccak256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_k1_ecrecover_keccak256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_ecrecover_sha256_cost_base(&self) -> u64

pub fn ecdsa_k1_ecrecover_sha256_cost_base_as_option(&self) -> Option<u64>

pub fn ecdsa_k1_ecrecover_sha256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_k1_ecrecover_sha256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_ecrecover_sha256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_k1_ecrecover_sha256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_decompress_pubkey_cost_base(&self) -> u64

pub fn ecdsa_k1_decompress_pubkey_cost_base_as_option(&self) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_keccak256_cost_base(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_keccak256_cost_base_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_keccak256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_keccak256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_keccak256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_keccak256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_sha256_cost_base(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_sha256_cost_base_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_sha256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_sha256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_k1_secp256k1_verify_sha256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_k1_secp256k1_verify_sha256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_ecrecover_keccak256_cost_base(&self) -> u64

pub fn ecdsa_r1_ecrecover_keccak256_cost_base_as_option(&self) -> Option<u64>

pub fn ecdsa_r1_ecrecover_keccak256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_r1_ecrecover_keccak256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_ecrecover_keccak256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_r1_ecrecover_keccak256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_ecrecover_sha256_cost_base(&self) -> u64

pub fn ecdsa_r1_ecrecover_sha256_cost_base_as_option(&self) -> Option<u64>

pub fn ecdsa_r1_ecrecover_sha256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_r1_ecrecover_sha256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_ecrecover_sha256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_r1_ecrecover_sha256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_keccak256_cost_base(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_keccak256_cost_base_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_keccak256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_keccak256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_keccak256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_keccak256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_sha256_cost_base(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_sha256_cost_base_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_sha256_msg_cost_per_byte(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_sha256_msg_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecdsa_r1_secp256r1_verify_sha256_msg_cost_per_block(&self) -> u64

pub fn ecdsa_r1_secp256r1_verify_sha256_msg_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ecvrf_ecvrf_verify_cost_base(&self) -> u64

pub fn ecvrf_ecvrf_verify_cost_base_as_option(&self) -> Option<u64>

pub fn ecvrf_ecvrf_verify_alpha_string_cost_per_byte(&self) -> u64

pub fn ecvrf_ecvrf_verify_alpha_string_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn ecvrf_ecvrf_verify_alpha_string_cost_per_block(&self) -> u64

pub fn ecvrf_ecvrf_verify_alpha_string_cost_per_block_as_option( &self ) -> Option<u64>

pub fn ed25519_ed25519_verify_cost_base(&self) -> u64

pub fn ed25519_ed25519_verify_cost_base_as_option(&self) -> Option<u64>

pub fn ed25519_ed25519_verify_msg_cost_per_byte(&self) -> u64

pub fn ed25519_ed25519_verify_msg_cost_per_byte_as_option(&self) -> Option<u64>

pub fn ed25519_ed25519_verify_msg_cost_per_block(&self) -> u64

pub fn ed25519_ed25519_verify_msg_cost_per_block_as_option(&self) -> Option<u64>

pub fn groth16_prepare_verifying_key_bls12381_cost_base(&self) -> u64

pub fn groth16_prepare_verifying_key_bls12381_cost_base_as_option( &self ) -> Option<u64>

pub fn groth16_prepare_verifying_key_bn254_cost_base(&self) -> u64

pub fn groth16_prepare_verifying_key_bn254_cost_base_as_option( &self ) -> Option<u64>

pub fn groth16_verify_groth16_proof_internal_bls12381_cost_base(&self) -> u64

pub fn groth16_verify_groth16_proof_internal_bls12381_cost_base_as_option( &self ) -> Option<u64>

pub fn groth16_verify_groth16_proof_internal_bls12381_cost_per_public_input( &self ) -> u64

pub fn groth16_verify_groth16_proof_internal_bls12381_cost_per_public_input_as_option( &self ) -> Option<u64>

pub fn groth16_verify_groth16_proof_internal_bn254_cost_base(&self) -> u64

pub fn groth16_verify_groth16_proof_internal_bn254_cost_base_as_option( &self ) -> Option<u64>

pub fn groth16_verify_groth16_proof_internal_bn254_cost_per_public_input( &self ) -> u64

pub fn groth16_verify_groth16_proof_internal_bn254_cost_per_public_input_as_option( &self ) -> Option<u64>

pub fn groth16_verify_groth16_proof_internal_public_input_cost_per_byte( &self ) -> u64

pub fn groth16_verify_groth16_proof_internal_public_input_cost_per_byte_as_option( &self ) -> Option<u64>

pub fn hash_blake2b256_cost_base(&self) -> u64

pub fn hash_blake2b256_cost_base_as_option(&self) -> Option<u64>

pub fn hash_blake2b256_data_cost_per_byte(&self) -> u64

pub fn hash_blake2b256_data_cost_per_byte_as_option(&self) -> Option<u64>

pub fn hash_blake2b256_data_cost_per_block(&self) -> u64

pub fn hash_blake2b256_data_cost_per_block_as_option(&self) -> Option<u64>

pub fn hash_keccak256_cost_base(&self) -> u64

pub fn hash_keccak256_cost_base_as_option(&self) -> Option<u64>

pub fn hash_keccak256_data_cost_per_byte(&self) -> u64

pub fn hash_keccak256_data_cost_per_byte_as_option(&self) -> Option<u64>

pub fn hash_keccak256_data_cost_per_block(&self) -> u64

pub fn hash_keccak256_data_cost_per_block_as_option(&self) -> Option<u64>

pub fn hmac_hmac_sha3_256_cost_base(&self) -> u64

pub fn hmac_hmac_sha3_256_cost_base_as_option(&self) -> Option<u64>

pub fn hmac_hmac_sha3_256_input_cost_per_byte(&self) -> u64

pub fn hmac_hmac_sha3_256_input_cost_per_byte_as_option(&self) -> Option<u64>

pub fn hmac_hmac_sha3_256_input_cost_per_block(&self) -> u64

pub fn hmac_hmac_sha3_256_input_cost_per_block_as_option(&self) -> Option<u64>

pub fn scoring_decision_mad_divisor(&self) -> f64

pub fn scoring_decision_mad_divisor_as_option(&self) -> Option<f64>

pub fn scoring_decision_cutoff_value(&self) -> f64

pub fn scoring_decision_cutoff_value_as_option(&self) -> Option<f64>

pub fn lookup_attr(&self, value: String) -> Option<ProtocolConfigValue>

Lookup a config attribute by its string representation

pub fn attr_map(&self) -> BTreeMap<String, Option<ProtocolConfigValue>>

Get a map of all config attribute from string representations

pub fn lookup_feature(&self, value: String) -> Option<bool>

Get the feature flags

pub fn feature_map(&self) -> BTreeMap<String, bool>

impl ProtocolConfig

pub fn check_package_upgrades_supported(&self) -> Result<(), Error>

pub fn package_upgrades_supported(&self) -> bool

pub fn check_commit_root_state_digest_supported(&self) -> bool

pub fn get_advance_epoch_start_time_in_safe_mode(&self) -> bool

pub fn loaded_child_objects_fixed(&self) -> bool

pub fn missing_type_is_compatibility_error(&self) -> bool

pub fn scoring_decision_with_validity_cutoff(&self) -> bool

pub fn narwhal_versioned_metadata(&self) -> bool

pub fn consensus_order_end_of_epoch_last(&self) -> bool

pub fn disallow_adding_abilities_on_upgrade(&self) -> bool

pub fn disable_invariant_violation_check_in_swap_loc(&self) -> bool

pub fn advance_to_highest_supported_protocol_version(&self) -> bool

pub fn ban_entry_init(&self) -> bool

pub fn package_digest_hash_module(&self) -> bool

pub fn disallow_change_struct_type_params_on_upgrade(&self) -> bool

pub fn no_extraneous_module_bytes(&self) -> bool

pub fn zklogin_auth(&self) -> bool

impl ProtocolConfig

pub fn get_for_version(version: ProtocolVersion, chain: Chain) -> Self

Get the value ProtocolConfig that are in effect during the given protocol version.

pub fn get_for_version_if_supported( version: ProtocolVersion, chain: Chain ) -> Option<Self>

Get the value ProtocolConfig that are in effect during the given protocol version. Or none if the version is not supported.

pub fn poison_get_for_min_version()

pub fn get_for_min_version() -> Self

Convenience to get the constants at the current minimum supported version. Mainly used by client code that may not yet be protocol-version aware.

pub fn get_for_max_version() -> Self

Convenience to get the constants at the current maximum supported version. Mainly used by genesis.

pub fn apply_overrides_for_testing( override_fn: impl Fn(ProtocolVersion, Self) -> Self + Send + 'static ) -> OverrideGuard

Override one or more settings in the config, for testing. This must be called at the beginning of the test, before get_for_(min|max)_version is called, since those functions cache their return value.

impl ProtocolConfig

pub fn set_max_function_definitions_for_testing(&mut self, m: u64)

pub fn set_buffer_stake_for_protocol_upgrade_bps_for_testing(&mut self, b: u64)

pub fn set_package_upgrades_for_testing(&mut self, val: bool)

pub fn set_advance_to_highest_supported_protocol_version_for_testing( &mut self, val: bool )

pub fn set_commit_root_state_digest_supported(&mut self, val: bool)

pub fn set_zklogin_auth(&mut self, val: bool)