Crate sui_protocol_config

Macros

check_limit

check_limit_by_meter

Used to check which limits were crossed if the TX is metered (not system tx) Args are: is_metered, value_to_check, metered_limit, unmetered_limit metered_limit is always less than or equal to unmetered_hard_limit

Structs

AliasedAddress

An aliased address.

Error

ExecutionTimeEstimateParams

OverrideGuard

ProtocolConfig

Constants that change the behavior of the protocol.

ProtocolConfigOptional

ProtocolVersion

Enums

Chain

ConsensusChoice

ConsensusNetwork

ConsensusTransactionOrdering

Ordering mechanism for transactions in one Narwhal consensus output.

LimitThresholdCrossed

Defines which limit got crossed. The value which crossed the limit and value of the limit crossed are embedded

PerObjectCongestionControlMode

ProtocolConfigValue

Functions

check_limit_in_range

Convenience function for comparing limit ranges V::MAX must be at >= U::MAX and T::MAX